Rename api to rpc
